refactor: 音周りと状態管理周りでファイル分けする
